LPD-84993 Toggle Product Versioning at instance scope
