+++ modules/test/playwright/pages/commerce/commerceInstanceSettingsPage.ts
@@ -6,15 +6,19 @@
 import {Locator, Page} from '@playwright/test';
 
 import {GlobalMenuPage} from '../product-navigation-applications-menu/GlobalMenuPage';
+import {UIElementsPage} from '../uielements/UIElementsPage';
 
 export class CommerceInstanceSettingsPage {
 	readonly globalMenuPage: GlobalMenuPage;
 	readonly catalogLink: Locator;
 	readonly checkboxPlacedOrders: (checkboxName: string) => Locator;
+	readonly editConfigurationSubmitButton: Locator;
+	readonly enabledButton: Locator;
 	readonly page: Page;
 	readonly productOptionMenuItem: Locator;
 	readonly showUnselectableOptionsCheckbox: Locator;
 	readonly submitConfigurationButton: Locator;
+	private uiElementsPage;
 
 	constructor(page: Page) {
 		this.globalMenuPage = new GlobalMenuPage(page);
@@ -24,6 +28,10 @@ export class CommerceInstanceSettingsPage {
 		});
 		this.checkboxPlacedOrders = (checkboxName) =>
 			page.getByLabel(checkboxName, {exact: true});
+		this.editConfigurationSubmitButton = page
+			.getByRole('button', {name: 'Save'})
+			.or(page.getByRole('button', {name: 'Update'}));
+		this.enabledButton = page.getByLabel('enabled');
 		this.page = page;
 		this.productOptionMenuItem = page.getByRole('menuitem', {
 			name: 'Product Options',
@@ -34,6 +42,7 @@ export class CommerceInstanceSettingsPage {
 		this.submitConfigurationButton = page.getByTestId(
 			'submitConfiguration'
 		);
+		this.uiElementsPage = new UIElementsPage(page);
 	}
 
 	async goto() {
@@ -54,6 +63,14 @@ export class CommerceInstanceSettingsPage {
 				name: configurationName,
 			})
 			.click();
+	}
+
+	async toggleProductVersioning() {
+		await this.globalMenuPage.goToHome();
+		await this.goToInstanceSetting('Catalog', 'Product Versioning');
+		await this.enabledButton.click();
+		await this.editConfigurationSubmitButton.click();
+		await this.uiElementsPage.anySuccessAlert.waitFor({state: 'visible'});
 	}
 
 	async toggleShowUnselectableOptions(check: boolean) {
